Always force use of correct ECC curve for ECDSA hybrids that specify a particular one
This change coerces the curve bit length in all cases in ssh-keygen, even when a scheme name is provided.

Calling ssh-keygen with the short name version of an ECDSA hybrid that calls for a particular curve that isn't P-256 (p384-rainbowiiicclassic and p521-rainbowvcclassic at present) would actually cause a key pair on P-256 to be generated as the classical part. Using the short name is very common.

This key would then not be recognized when loaded, because the EC nids would not match with the keytypes table.