Add liboqs-backed implementation of sntrup761x25519-sha512@openssh.com
`sntrup761x25519-sha512@openssh.com` is a PQ key exchange algorithm available in upstream OpenSSH. This implements that algorithm using the sntrup761 implementation within liboqs. Using this, we are able to demonstrate interop with upstream OpenSSH for this specific algorithm and more generally the construction of the hybrid combiner. While sntrup761 has fallen out of fashion, we have received [reassurances](https://github.com/open-quantum-safe/liboqs/pull/1328#issuecomment-1354111314) from upstream OpenSSH that it will be supported until at least 2027.

Support for the algorithm is added by adding the necessary entry to `generate.yml` and generating the relevant code snippets. In addition, we had to address the conflicting implementation from upstream by disabling the preprocessor macro and/or removing the conflicting code segment.

Testing

Updated `try_connection.py` to accept a specific KEX and/or SIG to test. This allows us to test arbitrary algorithms against arbitrary SSH implementations including the upstream OpenSSH. It was necessary to add a classical signature to the test so that there would be a supported signature in upstream OpenSSH that OQS could test against. In a follow-up PR I plan to add a CI job to do this, but am running short on time this week.

Related to Issue #163

Signed-off-by: Gerardo Ravago <gcr@amazon.com>