fix build failures when upgrading to 9.9P2
Additional clean-up from merge conflict resolution. The kex
defintions were resolved to use the existing implementation of
'KEX_ML_KEM_768_X25519_SHA256' instead of what openssh enabled
in V_9_9_P1. The 'kex-names.c' file had to be updated to reflect
this change. The ssh-rsa.c file was modified to remove a old call
to 'oqs_utils_is_rsa_hybrid' as the key parameter is no longer
available in that function.